Check whether deps are satisifed on service start, and fail if we're called as 'install_initd'. (#474223)
The LSB specifies that service add should fail if dependencies are not met,
and service removal should fail if things still rely on the service. This
implements the former.

This is tied only to LSB invocation, as this is bad behavior for the general
case, and downright dangerous for the removal case.