Fix more LSB bugs. (#474223)
When removing services in LSB mode, we want to not remove if dependencies
won't be met. However, we can't just check that the dependency exists on
some script in /etc/init.d - we must also check that that script is actually
set up to start/stop in a way that would cause the dependency to hit.