Only apply changes to start/stop priorities when the dependency is 'active'. (#474223)
When checking for start/stop priorities being handled, only frob the
priorities for start dependencies that are also configured to start,
and vice versa for stop dependencies.

Before we were applying the dependencies even if the providing service
wasn't configured, or wasn't configured to start. That isn't correct.

This is a behavior change, and could cause weirdness.