systemd: forward chkconfig on/off to systemctl enable/disable
Bill,

I am assuming that you are the maintainer of chkconfig? Here's a patch
adding the glue code for chkconfig to redirect things to systemctl, if
applicable. If systemd is not around this patch doesn't do anything.

The chkconfig sources had trailing whitespaces, this patch cleans that
up as side effect. I hope that's ok to merge. If not, I can send you a
version without whitespace changes. (Though I must say that cleaning up
whitespace is a great thing in order to ensure that it's easy to
generate clean patches in future).

This patch fowards chkconfig on/off as well as the chkconfig with only a
service name to their systemctl counterparts.

This also adds a message to chkconfig --list to make clear that the data
shown is incomplete and unreliable.

The SysV specific commands (such as runlevel or priority related
operations) are left unmodified since they cannot be translated properly
and it's probably safest to just leave them working as they are so that
existing package scripts just continue to work.