leveldb: do not try to close() fd -1
Calling chkconfig with a non-existing service name previously caused
close() to be called with fd -1 to the effect that errno was reset to
EBADF which was then reported to the user as error cause for the failed
operation. This patch makes this invocation go away so that errno is
left untouched and the user is not mislead with an error message that
doesn't actually explain what really failed.