systemd: don't redirect to "systemctl is-enabled" if runlevel is specified
If the user passes --levels then skip the redirection of chkconfig to
systemctl is-enable and instead answer exactly what the user asked for.

This is useful for implementing upgrades properly, since in that case we
need to query whether the SysV service was enabled to apply it to the
systemd service.