Use the newly added currentLevels for dependency checking.
This allows us to fix bug #693202/#771455 properly - instead of setting the
state on the filesystem before calling frobDependencies(), which was always
a hack, we merely adjust currentLevels to reflect the desired target state when
adding/enabling/disabling services, thereby only setting the final state
once.

Note: because of this, when entering frobDependencies, we always have to
adjust the read-from-the-filesystem state of the target service to match
the target service being passed in.

As a side-effect, isSimilarlyConfigured() becomes much shorter now that we
don't have to do level-by-level comparisons.