alternatives: add --keep-foreign
Right now when we are switching between alternatives or removing them,
we check if the facility (the path what user is using) is a symlink and
only refuse to remove it if it is an binary.

This commit add options --keep-foreign. With this option, alternatives
will check if the facility does not point to some other location than
altdir.