Fix systemd dependency
systemd is used during the build only to get systemunitdir from
systemd.pc. That pkgconfig file is now part of systemd, and
systemd-devel does not provide that, even indirectly.  While beakerlib
ends up pulling in systemd, that should not be relied upon, and does not
help for RHEL/ELN where beakerlib is now disabled.