Ignore alternatives that are binary-identical to existing ones
In https://bugzilla.redhat.com/show_bug.cgi?id=2363937 we found
a problem that is ultimately triggered by alternatives configs
having multiple entries that point to the same binary, or the
same *effective* binary after /usr and /sbin merges (which is
what streq_bin handles). I can't see a reason why we'd ever want
to support this as a real thing, so when reading the config,
let's just skip ingesting any alternative whose leader target is
the same effective binary as an alternative we've already read.

Signed-off-by: Adam Williamson <awilliam@redhat.com>